refactor: refactor messages
