+++ .eslintrc.json
@@ -18,7 +18,7 @@
         "plugin:@typescript-eslint/recommended"
     ],
     "rules": { 
-        "no-console": 1,
+        "no-console": 0,
         "indent": ["warn", 4],
         "eol-last": ["warn", "always"]
     }
