+++ src/app/requestListener.ts
@@ -1,7 +1,7 @@
 import { IncomingMessage, ServerResponse } from 'node:http';
 
 import { apiController } from './utils/apiController';
-import { StatusCodes, ErrorMessages } from './utils/types';
+import { StatusCodes, ErrorMessages } from './utils/messages';
 
 const requestListener = (req: IncomingMessage, res: ServerResponse) => {
     try {
